requirements.txt
absl-py==1.4.0
# alabaster==0.7.13
# astor==0.8.1
# asttokens==2.2.1
# astunparse==1.6.3
# autograd==1.6.2
# autograd-gamma==0.5.0
# Babel==2.12.1
# backcall==0.2.0
# boto3==1.34.6
# botocore==1.34.6
# cachetools==5.3.1
# certifi==2023.5.7
# cffi==1.15.1
# charset-normalizer==3.1.0
# click==8.1.6
# cloudpickle==2.2.1
# cmake==3.26.4
colored-traceback==0.3.0
comm==0.1.3
# contextlib2==21.6.0
# contourpy==1.1.0
# cycler==0.11.0
# Cython==3.0.0a10
D4RL==1.1
# debugpy==1.6.7
# decorator==5.1.1
# dill==0.3.7
dm-control==1.0.14
dm-env==1.6
dm-tree==0.1.8
docutils==0.18.1
exceptiongroup==1.1.2
executing==1.2.0
# Farama-Notifications==0.0.4
# fasteners==0.18
# filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
formulaic==0.6.6
future==0.18.3
gast==0.3.3
gin-config==0.5.0
glfw==2.6.2
google-api-core==2.11.1
# google-auth==2.21.0
# google-auth-oauthlib==1.0.0
# google-cloud-core==2.3.3
# google-cloud-storage==2.10.0
# google-crc32c==1.5.0
# google-pasta==0.2.0
# google-resumable-media==2.5.0
# googleapis-common-protos==1.59.1
graphlib-backport==1.0.3
grpcio==1.57.0
gym==0.23.0
gym-notices==0.0.8
gymnasium==0.28.1
h5py==2.10.0
idna==3.4
ImageHash==4.3.1
imageio==2.31.1
imagesize==1.4.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
interface-meta==1.3.0
ipdb==0.13.13
ipykernel==6.23.3
ipython==8.12.2
jax==0.4.13
jax-jumpy==1.0.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
labmaze==1.0.6
lazy_loader==0.3
libclang==16.0.6
lifelines==0.27.8
lit==16.0.6
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
# mjrl==1.0.0
ml-collections==0.1.1
ml-dtypes==0.2.0
mpmath==1.3.0
mujoco==2.3.7
mujoco-py==2.1.2.14
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
numpydoc==1.5.0
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cuda-cupti-cu11==11.7.101
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.2.10.91
# nvidia-cusolver-cu11==11.4.0.1
# nvidia-cusparse-cu11==11.7.4.91
# nvidia-nccl-cu11==2.14.3
# nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==10.0.0
pip==23.2.1
platformdirs==3.8.0
pluggy==1.2.0
prefetch-generator==1.0.3
prompt-toolkit==3.0.38
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.2.5
pycparser==2.21
pygame==2.1.3
Pygments==2.15.1
PyOpenGL==3.1.7
pyparsing==3.1.0
pytest==7.4.0
pytest-instafail==0.3.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
quadprog==0.1.11
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
setuptools==68.0.0
six==1.15.0
snowballstemmer==2.2.0
# Sphinx==6.2.1
# sphinx-rtd-theme==1.2.2
# sphinxcontrib-applehelp==1.0.4
# sphinxcontrib-devhelp==1.0.2
# sphinxcontrib-htmlhelp==2.0.1
# sphinxcontrib-jquery==4.1
# sphinxcontrib-jsmath==1.0.1
# sphinxcontrib-qthelp==1.0.3
# sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-addons==0.21.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-probability==0.20.1
termcolor==1.1.0
tf-agents==0.17.0
threadpoolctl==3.2.0
tick==0.7.0.1
tifffile==2023.7.10
tomli==2.0.1
torch==2.0.1
torch-ema==0.3
torchinfo==1.8.0
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
# traitlets==5.9.0
# triton==2.0.0
# typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
wget==3.2
wheel==0.38.4
wrapt==1.12.1
zipfile37==0.1.3
zipp==3.15.0
